fix: break datasets barrel circular imports causing Turbopack SSR 500
Next 16 builds with Turbopack by default, which is strict about
circular-dependency init order. Leaf controls (contextual-layers/controls,
widget-controls, contextual/card) imported INFO/DOWNLOAD/LAYERS from the
@/containers/datasets barrel, which imports the widgets that render those
controls — a cycle. Webpack tolerated the runtime-only access; Turbopack
threw 'Cannot access w before initialization' during module eval, 500ing
every route.

Split the per-slug registries (LAYERS, MAP_LEGENDS, INFO, DOWNLOAD,
BASEMAPS) into datasets/registries.ts, which imports only content modules
(info/layer/legend/download), and re-export them from the barrel. Leaf
consumers now import from registries.ts, severing the cycle.
