+++ src/components/widget-controls/index.tsx
@@ -1,4 +1,4 @@
-import { DOWNLOAD, INFO, LAYERS } from '@/containers/datasets';
+import { DOWNLOAD, INFO, LAYERS } from '@/containers/datasets/registries';
 
 import type { WidgetSlugType } from 'types/widget';
